refactor(tests): modulariza os testes de cada view d
