+++ api/users/tests/__init__.py
@@ -1,0 +1,3 @@
+from users.tests.register import *
+from users.tests.logout import *
+from users.tests.login import *
